pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "joshua7"
version = "0.2.0"
description = "Pre-publication AI content validation engine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Content Shield (Iron Forge Studios)"},
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "typer[all]>=0.12.0",
    "pyyaml>=6.0",
    "textstat>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[project.scripts]
joshua7 = "joshua7.cli.main:app"

[tool.setuptools.packages.find]
include = ["joshua7*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
